+++ bika/lims/browser/analyses/view.py
@@ -1020,17 +1020,12 @@ class AnalysesView(BikaListingView):
             # We want this icon to only appear if the context is an AR
             return
 
-        if analysis_brain.worksheetanalysis_review_state != 'assigned':
-            # No need to go further. This analysis is not assigned to any WS
-            return
-
         analysis_obj = self.get_object(analysis_brain)
-        worksheet = analysis_obj.getBackReferences('WorksheetAnalysis')
+        worksheet = analysis_obj.getWorksheet()
         if not worksheet:
             # No worksheet assigned. Do nothing
             return
 
-        worksheet = worksheet[0]
         title = t(_("Assigned to: ${worksheet_id}",
                     mapping={'worksheet_id': safe_unicode(worksheet.id)}))
         img = get_image('worksheet.png', title=title)