+++ bika/lims/catalog/analysis_catalog.py
@@ -22,7 +22,6 @@ CATALOG_ANALYSIS_LISTING = 'bika_analysis_catalog'
 # Defining the indexes for this catalog
 _indexes_dict = {
     'sortable_title': 'FieldIndex',
-    'worksheetanalysis_review_state': 'FieldIndex',
     'cancellation_state': 'FieldIndex',
     'getParentUID': 'FieldIndex',
     'getRequestUID': 'FieldIndex',
@@ -56,10 +55,10 @@ _indexes_dict = {
     'getPrioritySortkey': 'FieldIndex',
     'getAncestorsUIDs': 'KeywordIndex',
     'isSampleReceived': 'BooleanIndex',
+    'isWorksheetAssigned': 'BooleanIndex',
 }
 # Defining the columns for this catalog
 _columns_list = [
-    'worksheetanalysis_review_state',
     'getAttachmentUIDs',
     'getRequestID',
     'getReferenceAnalysesGroupID',