+++ bika/lims/exportimport/instruments/resultsimport.py
@@ -558,8 +558,7 @@ class AnalysisResultsImporter(Logger):
                     # And only if the filename of the attachment is unique in
                     # this worksheet.  Otherwise we will attempt to use
                     # existing attachment.
-                    wss = analysis.getBackReferences('WorksheetAnalysis')
-                    ws = wss[0] if wss else None
+                    ws = analysis.getWorksheet()
                     if ws:
                         if ws.getId() not in attachments:
                             fn = infile.filename
@@ -826,8 +825,8 @@ class AnalysisResultsImporter(Logger):
                 # The search has been made using the internal identifier
                 # from a Reference Analysis (id or uid). That is not usual.
                 an = refans[0].getObject()
-                wss = an.getBackReferences('WorksheetAnalysis')
-                if wss and len(wss) > 0:
+                worksheet = an.getWorksheet()
+                if worksheet:
                     # A regular QC test (assigned to a Worksheet)
                     return [an, ]
                 elif an.getInstrument():